split pymatviz.rdf into plotly and helpers modules
fix copy-paste error in structure_viz plotly+helpers module doc string
