+++ pymatviz/__init__.py
@@ -56,7 +56,7 @@ from pymatviz.ptable import (
     ptable_scatters,
 )
 from pymatviz.rainclouds import rainclouds
-from pymatviz.rdf import element_pair_rdfs, full_rdf
+from pymatviz.rdf.plotly import element_pair_rdfs, full_rdf
 from pymatviz.relevance import precision_recall_curve, roc_curve
 from pymatviz.sankey import sankey_from_2_df_cols
 from pymatviz.scatter import (
