+++ pymatviz/rdf/__init__.py
@@ -1,0 +1,5 @@
+"""2D and 3D plots of Structures."""
+
+from __future__ import annotations
+
+from pymatviz.rdf.plotly import element_pair_rdfs, full_rdf
