+++ pymatviz/rdf/plotly.py
@@ -14,140 +14,19 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-import numpy as np
 import plotly
 import plotly.graph_objects as go
 from plotly.subplots import make_subplots
-from pymatgen.core import Structure
-from pymatgen.optimization.neighbors import find_points_in_spheres
-from scipy.signal import find_peaks
+
+from pymatviz.rdf.helpers import calculate_rdf, normalize_structures
 
 
 if TYPE_CHECKING:
     from collections.abc import Sequence
     from typing import Any
 
-    from pymatgen.util.typing import PbcLike
-
-
-def normalize_structures(
-    structures: Structure | Sequence[Structure] | dict[str, Structure],
-) -> dict[str, Structure]:
-    """Normalize input to a dictionary of structures."""
-    if not structures:
-        raise ValueError("No structures provided")
-
-    if isinstance(structures, Structure):
-        return {"": structures}
-    if isinstance(structures, list | tuple) and all(
-        isinstance(s, Structure) for s in structures
-    ):
-        return {s.formula: s for s in structures}
-    if isinstance(structures, dict):
-        return structures
-    raise TypeError(f"Invalid input format for {structures=}")
-
-
-def calculate_rdf(
-    structure: Structure,
-    center_species: str | None = None,
-    neighbor_species: str | None = None,
-    cutoff: float = 15,
-    n_bins: int = 75,
-    pbc: PbcLike = (True, True, True),
-) -> tuple[np.ndarray, np.ndarray]:
-    """Calculate the radial distribution function (RDF) for a given structure.
-
-    If center_species and neighbor_species are provided, calculates the partial RDF
-    for the specified element pair. Otherwise, calculates the full RDF.
-
-    Args:
-        structure (Structure): A pymatgen Structure object.
-        center_species (str, optional): Symbol of the central species. If None, all
-            species are considered.
-        neighbor_species (str, optional): Symbol of the neighbor species. If None, all
-            species are considered.
-        cutoff (float, optional): Maximum distance for RDF calculation. Default is 15 Å.
-        n_bins (int, optional): Number of bins for RDF calculation.
-            Default is 75.
-        pbc (tuple[int, int, int], optional): Periodic boundary conditions as any
-            3-tuple of 0s/1s. Defaults to (1, 1, 1).
-
-    Returns:
-        tuple[np.ndarray, np.ndarray]: Arrays of (radii, g(r)) values.
-    """
-    bin_size = cutoff / n_bins
-    radii = np.linspace(0, cutoff, n_bins + 1)[1:]
-    rdf = np.zeros_like(radii)
-
-    # Get indices of center and neighbor species
-    if center_species:
-        center_indices = [
-            i
-            for i, site in enumerate(structure)
-            if site.specie.symbol == center_species
-        ]
-    else:
-        center_indices = list(range(len(structure)))
-
-    if neighbor_species:
-        neighbor_indices = [
-            i
-            for i, site in enumerate(structure)
-            if site.specie.symbol == neighbor_species
-        ]
-    else:
-        neighbor_indices = list(range(len(structure)))
-
-    # If there are no center atoms or neighbor atoms, return an empty RDF
-    if not center_indices or not neighbor_indices:
-        return radii, rdf  # Return zeros if no centers or neighbors
-
-    center_neighbors = find_points_in_spheres(
-        all_coords=structure.cart_coords,
-        center_coords=structure.cart_coords[center_indices],
-        r=cutoff,
-        # Convert bools to ints (needed for cython code)
-        pbc=np.array([*map(int, pbc)]),
-        lattice=structure.lattice.matrix,
-    )
-
-    # Filter distances for the specific neighbor species and bin them
-    neighbor_set = set(neighbor_indices)
-    for idx1, idx2, _, dist in zip(*center_neighbors, strict=True):
-        if idx2 in neighbor_set and center_indices[idx1] != idx2 and 0 < dist < cutoff:
-            bin_index = min(int(dist / bin_size), n_bins - 1)
-            rdf[bin_index] += 1
-
-    # Normalize RDF by the number of center-neighbor pairs and shell volumes
-    n_center = len(center_indices)
-    n_neighbor = len(neighbor_indices)
-    if center_species == neighbor_species:
-        normalization = n_center * (n_neighbor - 1)  # Exclude self-interactions
-    else:
-        normalization = n_center * n_neighbor
-
-    if normalization == 0:  # Avoid division by zero
-        return radii, np.zeros_like(radii)
-
-    # Spherical shell volume = surface area (4πr²) times thickness (bin_size)
-    rdf /= normalization
-    shell_volumes = 4 * np.pi * radii**2 * bin_size
-    rdf /= shell_volumes / structure.volume
-
-    return radii, rdf
-
-
-def find_last_significant_peak(
-    radii: np.ndarray, rdf: np.ndarray, prominence: float = 0.1
-) -> float:
-    """Find the position of the last significant peak in the RDF."""
-    peaks, properties = find_peaks(rdf, prominence=prominence, distance=5)
-    if peaks.size > 0:
-        # Sort peaks by prominence and select the last significant one
-        sorted_peaks = peaks[np.argsort(properties["prominences"])]
-        return radii[sorted_peaks[-1]]
-    return radii[-1]
+    import numpy as np
+    from pymatgen.core import Structure
 
 
 def element_pair_rdfs(
